// frontend/src/api/generated/model/adminControllerGetAllStartupsStatus.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Inside Line API
 * Inside Line API Documentation
 * OpenAPI spec version: 2.0
 */

export type AdminControllerGetAllStartupsStatus = typeof AdminControllerGetAllStartupsStatus[keyof typeof AdminControllerGetAllStartupsStatus];


export const AdminControllerGetAllStartupsStatus = {
  draft: 'draft',
  submitted: 'submitted',
  analyzing: 'analyzing',
  pending_review: 'pending_review',
  approved: 'approved',
  rejected: 'rejected',
} as const;